+++ src/Strain.js
@@ -13,6 +13,7 @@
 const URI = require('uri-js');
 const YAML = require('yaml');
 const { YAMLMap, Pair } = require('yaml/types');
+const log = require('@adobe/helix-log');
 
 const GitUrl = require('./GitUrl.js');
 const Origin = require('./Origin.js');
@@ -21,7 +22,12 @@ const Performance = require('./Performance.js');
 const Redirect = require('./Redirect.js');
 const Condition = require('../src/Condition.js');
 const utils = require('./utils.js');
-const log = require('./log.js');
+
+/**
+ * Flags indicating whether deprecation warning were shown.
+ */
+let urlOverridesCondition;
+let urlIsDeprecated;
 
 /**
  * Strain
@@ -51,10 +57,12 @@ class Strain {
     this._condition = new Condition(cfg.condition || '');
 
     if (cfg.url) {
-      if (cfg.condition) {
+      if (cfg.condition && !urlOverridesCondition) {
         log.warn('Property url overrides property condition, use just a condition instead.');
-      } else {
-        log.info('Property url is deprecated, use a condition instead.');
+        urlOverridesCondition = 1;
+      } else if (!urlIsDeprecated) {
+        log.info(`Property url is deprecated, use a condition instead:\ncondition:\n  url: ${cfg.url}`);
+        urlIsDeprecated = 1;
       }
     }