+++ src/index.js
@@ -11,22 +11,24 @@
  */
 const GitUrl = require('./GitUrl.js');
 const HelixConfig = require('./HelixConfig.js');
+const IndexConfig = require('./IndexConfig');
+const MountConfig = require('./MountConfig');
 const Strain = require('./Strain.js');
 const utils = require('./utils.js');
 const string = require('./string.js');
 const dom = require('./dom.js');
-const log = require('./log.js');
-const Logger = require('./Logger.js');
 const Async = require('./async.js');
+const MarkupConfig = require('./MarkupConfig');
 
 module.exports = {
   GitUrl,
   HelixConfig,
+  IndexConfig,
+  MountConfig,
   Strain,
   string,
   dom,
   utils,
-  log,
-  Logger,
+  MarkupConfig,
   async: Async,
 };